test(torghut): split autonomy lane helper coverage
